#400: refactoring
